Just enough to get incorrectly-colored graphics onscreen
This is really just a checkpoint commit immediately after the first
successful rendering of the title screen and demo playback.

This currently sends the pixmap to the screen but not the colormap, so the
colors are all wrong (using the RISCovite default palette) but it's
encouraging to see it rendering something plausible to the screen
nonetheless.