Audio volume and stereo panning
At this point the sound effect playback is mostly working, but there is
some glitching that I've not yet been able to explain when lots of samples
play back in quick succession. It seems like there might be a
synchronization problem where channels are getting activated and
deactivated in an incorrect order relative to the other channel field
changes.