Only submit the part of the sound buffer that we've populated
This was part of the explanation for the audio glitches: an earlier version
of this code tried to submit fixed-sized audio buffer chunks, but I later
switched it to just fill up whatever buffer length the interrupt handler
was told it could safely generate but forgot to update the write_samples
call to actually take that into account.

There's still something not quite right here because some samples are
getting truncated, but at least we're no longer accidentally submitting
older audio data from previous iterations.